Change to original base order price
